handle errors better, finish submit form inputs
This allows the submit form to function now.

The errors handling was based around the handling strategy in the uisvc
where a json result is produced in the body for each 500 error. This
allows the client to bypass a lot of typical typing crap we don't want
to do to always echo the same thing, but apparently the JS swagger client
doesn't like operating that way (With any data on a 500), so we hacked
around that by adding a response text check and JSON unmarshal. It seems
to work and hopefully the biggest problem with it is null pointer
mishaps.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>