submission ID support in UI
This enables the submission ID as the primary source of navigation for
the UI. It allows the user to see at-a-glance testing states of full
suites, instead of having to derive it from the task list.

Cancellations only report a failure for submissions; not a canceled
state. This is something I hope to resolve soon but the work involved is
pretty large so I'm going to hold off for now.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>